slight refactoring
