+++ lib/config/service_locator/service_locator.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 import 'package:isar/isar.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:todo_list/core/services/device_info_service.dart';
+import 'package:todo_list/core/services/device_info/device_info_service.dart';
 import 'package:todo_list/core/services/preferences/preferences_service/preferences_service.dart';
 import 'package:todo_list/features/todo/data/dto/local/local_todo_dto.dart';
 import 'package:todo_list/features/todo/data/repositories/todo_repository_impl.dart';
@@ -11,11 +11,11 @@ import 'package:todo_list/features/todo/data/sources/remote/remote_source/remote
 import 'package:todo_list/features/todo/domain/repository/todo_repository.dart';
 
 Future<void> initDependecies() async {
-  GetIt.I.registerSingletonAsync<PreferencesService>(
-    () => PreferencesService().init(),
-  );
   GetIt.I.registerSingletonAsync<DeviceInfoService>(
     () => DeviceInfoService().init(),
+  );
+  GetIt.I.registerSingletonAsync<PreferencesService>(
+    () => PreferencesService().init(),
   );
   GetIt.I.registerSingletonAsync<TodoRepository>(
     () async {
