+++ lib/features/todo/presentation/todo_all/widgets/list/components/fast_todo_creation_tile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:todo_list/core/extensions/theme_extension.dart';
-import 'package:todo_list/core/services/device_info_service.dart';
+import 'package:todo_list/core/services/device_info/device_info_service.dart';
 import 'package:todo_list/core/ui/layout/app_shimmer.dart';
 import 'package:todo_list/core/ui/layout/custom_card.dart';
 import 'package:todo_list/core/ui/widget/custom_icon_button.dart';
