+++ lib/features/todo/presentation/todo_all/widgets/list/components/todo_tile/todo_tile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:todo_list/config/logger/logger.dart';
-import 'package:todo_list/core/services/device_info_service.dart';
+import 'package:todo_list/core/services/device_info/device_info_service.dart';
 import 'package:todo_list/core/ui/layout/custom_card.dart';
 import 'package:todo_list/features/todo/domain/state_management/todo_list_bloc/todo_list_bloc.dart';
 import 'package:todo_list/features/todo/domain/state_management/todo_list_bloc/todo_list_event.dart';
