+++ lib/features/todo/presentation/todo_single/widgets/buttons/todo_save_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
 import 'package:todo_list/core/extensions/theme_extension.dart';
-import 'package:todo_list/core/services/device_info_service.dart';
+import 'package:todo_list/core/services/device_info/device_info_service.dart';
 import 'package:todo_list/core/ui/layout/custom_button_base.dart';
 import 'package:todo_list/config/l10n/generated/l10n.dart';
 import 'package:uuid/uuid.dart';
